+++ web-console/src/index.html
@@ -10,7 +10,7 @@
   <body>
     <div class="console-background console-background--special" aria-hidden="true">
       <div class="console-background-grid">
-        <!-- 9 宫格使用同一张拼图，单元格通过 CSS background-position 显示对应切片 -->
+        <!-- 9 宫格使用同一张正方形拼图；网格保持正方形，background-size 300% 等比切格不变形 -->
         <div class="console-background-grid-cell"></div>
         <div class="console-background-grid-cell"></div>
         <div class="console-background-grid-cell"></div>
@@ -208,11 +208,19 @@
               <p class="section-kicker">TASK BOARD / LOCAL API</p>
               <h2 id="todo-heading">任务清单</h2>
             </div>
-            <button id="todo-refresh" class="secondary" type="button">刷新</button>
+            <div class="todo-page-actions">
+              <button id="todo-refresh" class="secondary" type="button">刷新</button>
+              <button id="todo-create-open" type="button">创建 Todo</button>
+            </div>
           </header>
           <div class="todo-toolbar">
             <label>状态<select id="todo-status-filter"><option value="all">全部</option><option value="pending">待处理</option><option value="completed">已完成</option></select></label>
             <label class="todo-search">关键词<input id="todo-keyword-filter" type="search" placeholder="搜索标题或内容"></label>
+            <button id="todo-advanced-toggle" type="button" class="secondary" aria-expanded="false">高级筛选</button>
+            <button id="todo-filter-submit" type="button">应用筛选</button>
+            <button id="todo-filter-reset" type="button" class="secondary">重置</button>
+          </div>
+          <div id="todo-advanced-filter" class="todo-advanced-filter" hidden>
             <label>时间<select id="todo-time-filter"><option value="all">全部</option><option value="overdue">已逾期</option><option value="no_due_date">无截止日期</option></select></label>
             <label>重复<select id="todo-recurring-filter"><option value="all">全部</option><option value="true">仅重复</option><option value="false">非重复</option></select></label>
             <label>目标<select id="todo-target-filter"><option value="">全部目标</option></select></label>
@@ -222,26 +230,32 @@
             <label>范围<select id="todo-scope-filter"><option value="">全部</option><option value="private">私聊</option><option value="group">群聊</option></select></label>
             <label>日期起<input id="todo-date-start" type="date"></label>
             <label>日期止<input id="todo-date-end" type="date"></label>
-            <button id="todo-filter-submit" type="button">应用筛选</button>
           </div>
-          <form id="todo-create-form" class="todo-create-form">
-            <label>标题<input id="todo-create-title" name="title" required maxlength="500" placeholder="例如：整理部署日志"></label>
-            <label>目标<select id="todo-create-target" name="target_ref" required><option value="">加载目标中…</option></select></label>
-            <label>截止日期<input id="todo-create-due-date" name="due_date" type="date"></label>
-            <label>截止时间<input id="todo-create-due-at" name="due_at" placeholder="YYYY-MM-DD HH:MM:SS"></label>
-            <label>提醒时间<input id="todo-create-reminder-at" name="reminder_at" placeholder="YYYY-MM-DD HH:MM:SS"></label>
-            <label>时间精度<select id="todo-create-time-precision"><option value="none">无</option><option value="date">日期</option><option value="date_time">日期时间</option></select></label>
-            <label>重复类型<select id="todo-create-recurrence-kind"><option value="none">不重复</option><option value="every_n_days">间隔重复</option></select></label>
-            <label>重复间隔<input id="todo-create-recurrence-interval" type="number" min="1" placeholder="可选"></label>
-            <label>重复单位<select id="todo-create-recurrence-unit"><option value="day">天</option><option value="week">周</option><option value="month">月</option></select></label>
-            <label class="todo-create-detail">详情<textarea id="todo-create-detail" name="detail" rows="2" placeholder="可选"></textarea></label>
-            <button type="submit">创建 Todo</button>
-          </form>
           <p id="todo-result" class="status-message" role="status" aria-live="polite"></p>
           <div id="todo-list" class="todo-list" aria-live="polite"><p class="hint">正在加载 Todo…</p></div>
           <div id="todo-pagination" class="todo-pagination"></div>
         </section>
       </div>
+
+      <dialog id="todo-create-dialog" class="todo-create-dialog">
+        <form id="todo-create-form" class="todo-create-form" method="dialog">
+          <div class="todo-create-heading">
+            <h3>创建 Todo</h3>
+            <button id="todo-create-close" type="button" class="secondary" aria-label="关闭">×</button>
+          </div>
+          <label>标题<input id="todo-create-title" name="title" required maxlength="500" placeholder="例如：整理部署日志"></label>
+          <label>目标<select id="todo-create-target" name="target_ref" required><option value="">加载目标中…</option></select></label>
+          <label>截止时间<input id="todo-create-due-at" name="due_at" type="datetime-local"></label>
+          <label>提醒时间<input id="todo-create-reminder-at" name="reminder_at" type="datetime-local"></label>
+          <label>时间精度<select id="todo-create-time-precision"><option value="none">无</option><option value="date_time">日期时间</option></select></label>
+          <label>重复类型<select id="todo-create-recurrence-kind"><option value="none">不重复</option><option value="every_n_days">间隔重复</option></select></label>
+          <label>重复间隔<input id="todo-create-recurrence-interval" type="number" min="1" placeholder="可选"></label>
+          <label>重复单位<select id="todo-create-recurrence-unit"><option value="day">天</option><option value="week">周</option><option value="month">月</option></select></label>
+          <label class="todo-create-detail">详情<textarea id="todo-create-detail" name="detail" rows="2" placeholder="可选"></textarea></label>
+          <p id="todo-create-error" class="error" role="alert"></p>
+          <button type="submit">创建</button>
+        </form>
+      </dialog>
 
       <div class="console-page" data-console-page="configuration" hidden>
       <header class="product-page-heading configuration-page-heading">
