--- web-console/src/views/configuration.ts
@@ -1,1596 +1,0 @@
-import {
-  ConsoleApiError,
-  fetchConfiguration,
-  requestRestart,
-  updateAgentConfiguration,
-  updateRuntimeConfiguration,
-  updateSecretConfiguration,
-  validateConfiguration,
-} from "../api.js";
-import { agentToolOptions, selectedAgentToolNames, type AgentToolOption } from "../agent-tools.js";
-import { togglePasswordReveal } from "../dom.js";
-import {
-  openCodeProviderChange,
-  readOpenCodeProviders,
-  renderOpenCodeProviders,
-  renderOpenCodeRouteHints,
-} from "../opencode-providers.js";
-import type { ConfigFieldSnapshot, ConfigurationSnapshot, UserFile, UserPreferences } from "../types.js";
-import type { ThemeController } from "../theme.js";
-import { renderThemeSelector } from "./theme-selector.js";
-import type { BackgroundController } from "../background.js";
-
-const FIELD_LABELS: Record<string, string> = {
-  "command.prefix": "聊天命令前缀",
-  "delivery.tts.provider": "语音 Provider",
-  "delivery.tts.qwen_api_key": "千问 TTS API Key",
-  "delivery.tts.qwen_base_url": "千问 TTS Base URL",
-  "delivery.tts.qwen_model": "千问 TTS 模型",
-  "delivery.tts.qwen_voice": "千问 TTS 音色",
-  "delivery.tts.request_timeout_seconds": "请求超时（秒）",
-  "delivery.tts.max_text_chars": "最大朗读字符数",
-  "provider.openai.base_url": "OpenAI Base URL",
-  "provider.openai.api_mode": "OpenAI API 模式",
-  "provider.openai.api_key": "OpenAI API Key",
-  "provider.deepseek.base_url": "DeepSeek Base URL",
-  "provider.deepseek.api_key": "DeepSeek API Key",
-  "provider.bigmodel.base_url": "BigModel Base URL",
-  "provider.bigmodel.api_key": "BigModel API Key",
-  "provider.gemini.base_url": "Gemini Base URL",
-  "provider.gemini.api_key": "Gemini API Key",
-  "provider.mimo.api_key": "MiMo API Key",
-  "provider.opencode.api_key": "OpenCode API Key",
-  "tools.web_search.tavily.api_key": "Tavily API Key",
-  "weather.qweather.api_key": "和风天气 API Key",
-  "weather.qweather.api_host": "QWeather API Host",
-  "weather.qweather.geo_host": "QWeather Geo Host",
-  "platform.qq_official.enabled": "QQ 官方入口",
-  "platform.qq_official.app_id": "QQ AppID",
-  "platform.qq_official.app_secret": "QQ AppSecret",
-  "platform.onebot11.enabled": "OneBot 11 入口",
-  "platform.onebot11.bind_host": "OneBot 绑定地址",
-  "platform.onebot11.bind_port": "OneBot 绑定端口",
-  "platform.onebot11.websocket_path": "OneBot WebSocket 路径",
-  "platform.onebot11.access_token": "OneBot Access Token",
-  "platform.wechat_service.enabled": "微信服务号入口",
-  "platform.wechat_service.token": "微信 Token",
-  "platform.wechat_service.app_id": "微信 AppID",
-  "platform.wechat_service.app_secret": "微信 AppSecret",
-  "platform.wechat_service.encryption_mode": "微信消息加密模式",
-  "platform.wechat_service.encoding_aes_key": "微信 EncodingAESKey",
-  "platform.wechat_service.bind_host": "微信回调监听地址",
-  "platform.wechat_service.bind_port": "微信回调监听端口",
-  "platform.wechat_service.callback_path": "微信回调路径",
-  "features.rss.enabled": "RSS",
-  "features.rss.translation_enabled": "RSS 翻译",
-  "features.memory.consolidation_enabled": "Memory 整理",
-  "features.memory.dream_enabled": "Session Dream",
-  "features.todo.daily_reminder_enabled": "Todo 每日提醒",
-  "features.todo.daily_reminder_time": "Todo 提醒时间",
-  "console.enabled": "Web 控制台",
-  "console.allowed_origins": "Web 控制台允许来源",
-  "console.trusted_proxy_ips": "Web 控制台可信代理 IP",
-  "console.secure_cookies": "Web 控制台安全 Cookie",
-  "bootstrap.listen_host": "Core 监听地址",
-  "bootstrap.listen_port": "Core 监听端口",
-  "bootstrap.database_file": "数据库文件",
-  "bootstrap.database_pool_size": "数据库连接池大小",
-  "bootstrap.runtime_config_file": "运行配置文件",
-  "bootstrap.master_key_file": "主密钥文件",
-  "bootstrap.agent_config_file": "Agent 配置文件",
-  "bootstrap.ops_config_file": "运维配置文件",
-};
-
-export type ConfigurationBusinessGroup =
-  | "models-providers"
-  | "model-routing"
-  | "online-tools"
-  | "memory-knowledge"
-  | "replies-voice"
-  | "platforms"
-  | "tasks-notifications"
-  | "system-security"
-  | "advanced";
-
-interface BusinessGroupDefinition {
-  id: ConfigurationBusinessGroup;
-  label: string;
-  description: string;
-}
-
-interface FieldSectionDefinition {
-  business: ConfigurationBusinessGroup;
-  label: string;
-  prefix: string;
-  description?: string;
-}
-
-const BUSINESS_GROUPS: readonly BusinessGroupDefinition[] = [
-  { id: "models-providers", label: "模型与供应商", description: "配置内置模型服务、自定义 Provider 及对应凭据。" },
-  { id: "model-routing", label: "模型路由", description: "为私聊、群聊、辅助任务和搜索场景设置候选模型链。" },
-  { id: "online-tools", label: "联网与工具", description: "管理联网搜索、天气服务、Tool Calling 与场景白名单。" },
-  { id: "memory-knowledge", label: "记忆与知识库", description: "管理记忆整理、Session Dream 和本地知识检索策略。" },
-  { id: "replies-voice", label: "回复与语音", description: "设置聊天命令入口和最终回复的语音能力。" },
-  { id: "platforms", label: "平台接入", description: "配置 QQ 官方、OneBot 11 与微信服务号入口。" },
-  { id: "tasks-notifications", label: "待办与通知", description: "管理 Todo 每日提醒、RSS 订阅和翻译开关。" },
-  { id: "system-security", label: "系统与安全", description: "查看控制台、安全边界、只读启动参数和界面偏好。" },
-  { id: "advanced", label: "高级兼容配置", description: "保留当前版本尚未归类的受管字段，不会在其他字段保存时删除。" },
-];
-
-const FIELD_SECTIONS: readonly FieldSectionDefinition[] = [
-  { business: "replies-voice", label: "命令交互", prefix: "command." },
-  {
-    business: "replies-voice",
-    label: "语音回复",
-    prefix: "delivery.tts.",
-    description: "修改后需要重启。Web 控制台只配置全局 TTS 能力；Provider 关闭或千问 API Key 未配置时，/语音 开启会被拒绝。私聊或群聊是否启用仍通过 /语音 开启、/语音 关闭管理。关闭 Provider 不会清除已保存的千问配置，也可以提前填写。",
-  },
-  { business: "models-providers", label: "内置模型服务", prefix: "provider." },
-  { business: "platforms", label: "QQ 官方入口", prefix: "platform.qq_official." },
-  { business: "platforms", label: "OneBot 11 入口", prefix: "platform.onebot11." },
-  { business: "platforms", label: "微信服务号入口", prefix: "platform.wechat_service." },
-  { business: "tasks-notifications", label: "RSS 订阅", prefix: "features.rss." },
-  { business: "memory-knowledge", label: "记忆策略", prefix: "features.memory." },
-  { business: "tasks-notifications", label: "Todo 提醒", prefix: "features.todo." },
-  { business: "online-tools", label: "联网搜索凭据", prefix: "tools.web_search." },
-  { business: "online-tools", label: "天气服务", prefix: "weather." },
-  { business: "system-security", label: "Web 控制台", prefix: "console." },
-  { business: "system-security", label: "基础运行（只读）", prefix: "bootstrap." },
-];
-
-const TTS_PROVIDER_OPTIONS: ReadonlyArray<readonly [string, string]> = [
-  ["disabled", "关闭"],
-  ["qwen", "千问"],
-];
-
-const TTS_NUMBER_RANGES: Readonly<Record<string, readonly [number, number]>> = {
-  "delivery.tts.request_timeout_seconds": [1, 120],
-  "delivery.tts.max_text_chars": [1, 600],
-};
-
-export function configFieldLabel(key: string): string {
-  return FIELD_LABELS[key] ?? key;
-}
-
-export function configFieldGroupLabel(key: string): string {
-  const business = configurationBusinessGroup(key);
-  return BUSINESS_GROUPS.find((group) => group.id === business)?.label ?? "高级兼容配置";
-}
-
-/** 配置键只在一个业务域出现；未知或后续新增字段进入高级兼容区，不会被静默遗漏。 */
-export function configurationBusinessGroup(key: string): ConfigurationBusinessGroup {
-  return FIELD_SECTIONS.find((section) => key.startsWith(section.prefix))?.business ?? "advanced";
-}
-
-/** 未识别的历史 Provider 必须原样保留，避免页面加载或保存时静默改成受支持值。 */
-export function ttsProviderOptions(currentValue: unknown): Array<[string, string]> {
-  const current = currentValue === null || currentValue === undefined ? "disabled" : String(currentValue);
-  const options = TTS_PROVIDER_OPTIONS.map(([value, label]): [string, string] => [value, label]);
-  if (!options.some(([value]) => value === current)) {
-    options.push([current, `${current}（当前自定义值）`]);
-  }
-  return options;
-}
-
-export function ttsNumberRange(key: string): readonly [number, number] | null {
-  return TTS_NUMBER_RANGES[key] ?? null;
-}
-
-/** TTS 范围字段必须先完整通过整数与边界校验，不能沿用普通整数的宽松 parseInt 语义。 */
-export function parseTtsNumberValue(key: string, rawValue: string): number {
-  const range = ttsNumberRange(key);
-  if (!range) throw new Error(`${configFieldLabel(key)}没有可用的页面输入范围`);
-  const value = rawValue.trim() === "" ? Number.NaN : Number(rawValue);
-  if (!Number.isFinite(value) || !Number.isInteger(value) || value < range[0] || value > range[1]) {
-    throw new Error(`${configFieldLabel(key)}必须是 ${range[0]} 到 ${range[1]} 之间的整数`);
-  }
-  return value;
-}
-
-const AGENT_ROUTE_LABELS: Record<string, string> = {
-  private_main: "私聊主路线",
-  group_main: "群聊主路线",
-  aux: "辅助任务路线",
-  private_search: "私聊搜索路线",
-  group_search: "群聊搜索路线",
-};
-
-export type AgentWebSearchBackend = "provider_native" | "tavily" | "disabled";
-
-export interface AgentWebSearchConfig {
-  backend: AgentWebSearchBackend;
-  maxResults: number;
-  searchDepth: "basic" | "advanced";
-  topic: "general" | "news" | "finance";
-  timeRange: "day" | "week" | "month" | "year" | null;
-  connectTimeoutSeconds: number;
-  firstResponseTimeoutSeconds: number;
-  totalTimeoutSeconds: number;
-  routes: Record<string, string>;
-}
-
-export type AutosaveScope = "public" | "secret" | "agent";
-export interface AutosaveBlurInput {
-  readonly scope: AutosaveScope;
-  readonly value: unknown;
-  readonly baseline: unknown;
-  readonly configured?: boolean;
-  readonly clearRequested?: boolean;
-}
-
-export interface UserDataController {
-  readonly preferences: UserPreferences;
-  readonly files: readonly UserFile[];
-  readonly updatePreferences: (patch: {
-    readonly customColors?: readonly string[];
-    readonly backgroundFileIds?: readonly string[];
-    readonly activeBackgroundFileId?: string | null;
-    readonly backgroundMode?: "default" | "special";
-    readonly kuliantnt?: boolean;
-  }) => Promise<UserPreferences>;
-  readonly uploadFile?: (file: File) => Promise<UserFile>;
-  readonly deleteFile?: (file: UserFile) => Promise<void>;
-}
-
-/** 仅测试使用：重置自动保存绑定与保存队列等模块级状态，避免多个测试文档互相泄漏。 */
-export function resetConfigurationStateForTests(): void {
-  current = null;
-  selectedBusinessGroup = "models-providers";
-  autosaveBound = false;
-  queuedFocusRestoreId = null;
-  secretSavedStates.clear();
-  saveQueue = Promise.resolve(null);
-}
-
-export function shouldAutosaveOnBlur(input: AutosaveBlurInput): boolean {
-  if (input.scope === "secret") {
-    if (input.clearRequested === true) return input.configured === true;
-    return typeof input.value === "string" && input.value.length > 0;
-  }
-  if (input.scope === "public" && (input.baseline === null || input.baseline === undefined) && isEmptyInputValue(input.value)) {
-    return false;
-  }
-  return JSON.stringify(input.value) !== JSON.stringify(input.baseline);
-}
-
-const DEFAULT_WEB_SEARCH_CONFIG: AgentWebSearchConfig = {
-  backend: "provider_native",
-  maxResults: 5,
-  searchDepth: "basic",
-  topic: "general",
-  timeRange: null,
-  connectTimeoutSeconds: 10,
-  firstResponseTimeoutSeconds: 30,
-  totalTimeoutSeconds: 60,
-  routes: {},
-};
-
-/** 只识别统一的 tools.web_search；旧顶层 search_routes 不参与页面读取。 */
-export function readAgentWebSearchConfig(documentValue: unknown): AgentWebSearchConfig {
-  const webSearch = record(record(record(documentValue).tools).web_search);
-  const routes = Object.fromEntries(
-    Object.entries(record(webSearch.routes))
-      .map(([name, value]) => [name, string(record(value).model)] as const)
-      .filter(([, model]) => model.length > 0),
-  );
-  const backend = string(webSearch.backend);
-  const searchDepth = string(webSearch.search_depth);
-  const topic = string(webSearch.topic);
-  const timeRange = string(webSearch.time_range);
-  return {
-    backend: isWebSearchBackend(backend) ? backend : DEFAULT_WEB_SEARCH_CONFIG.backend,
-    maxResults: positiveNumber(webSearch.max_results, DEFAULT_WEB_SEARCH_CONFIG.maxResults),
-    searchDepth: searchDepth === "advanced" ? "advanced" : "basic",
-    topic: topic === "news" || topic === "finance" ? topic : "general",
-    timeRange: isWebSearchTimeRange(timeRange) ? timeRange : null,
-    connectTimeoutSeconds: positiveNumber(webSearch.connect_timeout_seconds, DEFAULT_WEB_SEARCH_CONFIG.connectTimeoutSeconds),
-    firstResponseTimeoutSeconds: positiveNumber(webSearch.first_response_timeout_seconds, DEFAULT_WEB_SEARCH_CONFIG.firstResponseTimeoutSeconds),
-    totalTimeoutSeconds: positiveNumber(webSearch.total_timeout_seconds, DEFAULT_WEB_SEARCH_CONFIG.totalTimeoutSeconds),
-    routes,
-  };
-}
-
-export function webSearchConfigChange(config: AgentWebSearchConfig): Record<string, unknown> {
-  if (!Number.isInteger(config.maxResults) || config.maxResults < 1 || config.maxResults > 10) {
-    throw new Error("Tavily 结果数必须是 1 到 10 之间的整数");
-  }
-  for (const [label, value] of [
-    ["连接超时", config.connectTimeoutSeconds],
-    ["首响应超时", config.firstResponseTimeoutSeconds],
-    ["总超时", config.totalTimeoutSeconds],
-  ] as const) {
-    if (!Number.isInteger(value) || value < 1) throw new Error(`${label}必须是大于 0 的整数秒数`);
-  }
-  if (config.connectTimeoutSeconds > config.firstResponseTimeoutSeconds) {
-    throw new Error("连接超时不能大于首响应超时");
-  }
-  if (config.firstResponseTimeoutSeconds > config.totalTimeoutSeconds) {
-    throw new Error("首响应超时不能大于总超时");
-  }
-  return {
-    action: "set_web_search",
-    backend: config.backend,
-    max_results: config.maxResults,
-    search_depth: config.searchDepth,
-    topic: config.topic,
-    time_range: config.timeRange,
-    connect_timeout_seconds: config.connectTimeoutSeconds,
-    first_response_timeout_seconds: config.firstResponseTimeoutSeconds,
-    total_timeout_seconds: config.totalTimeoutSeconds,
-  };
-}
-
-export function tavilyCredentialNotice(backend: AgentWebSearchBackend, configured: boolean): string {
-  return backend === "tavily" && !configured
-    ? "已选择 Tavily，但 Tavily API Key 尚未配置。请先在“联网与工具”中保存 Key，重启后搜索才可用。"
-    : "";
-}
-
-export function webSearchRouteChanges(
-  savedRoutes: Record<string, string>,
-  formRoutes: Record<string, string>,
-): Array<Record<string, unknown>> {
-  const changes: Array<Record<string, unknown>> = [];
-  for (const name of ["private_search", "group_search"]) {
-    const model = (formRoutes[name] ?? "").trim();
-    // 后端切换只更新联网搜索参数；空输入或未改动路线都保留当前 agent.toml 内容。
-    if (model.length > 0 && model !== (savedRoutes[name] ?? "")) {
-      changes.push({ action: "set_search_route", name, model });
-    }
-  }
-  return changes;
-}
-
-let current: ConfigurationSnapshot | null = null;
-let currentThemeController: ThemeController | null = null;
-let currentBackgroundController: BackgroundController | null = null;
-let currentUserDataController: UserDataController | null = null;
-let toastTimer: number | undefined;
-let selectedBusinessGroup: ConfigurationBusinessGroup = "models-providers";
-let autosaveBound = false;
-let queuedFocusRestoreId: string | null = null;
-let saveQueue: Promise<ConfigurationSnapshot | null> = Promise.resolve(null);
-
-/** Secret 的“最后成功保存状态”。服务端不回传 secret 明文，前端只能靠本地记录判断脏状态，
- *  避免连续修改多个 Secret 时重复提交已经成功替换的旧 revision。 */
-interface SecretSavedState {
-  readonly value: string;
-  readonly clear: boolean;
-  readonly revision: string | null;
-}
-const secretSavedStates = new Map<string, SecretSavedState>();
-const EMPTY_EXCLUDED_KEYS: ReadonlySet<string> = new Set();
-
-export async function initializeConfiguration(
-  themeController: ThemeController,
-  backgroundController: BackgroundController,
-  userData: UserDataController | null = null,
-): Promise<void> {
-  currentThemeController = themeController;
-  currentBackgroundController = backgroundController;
-  currentUserDataController = userData;
-  // 每次（重新）初始化都清空跨登录会话残留的 Secret 已保存状态：服务端不回传明文，
-  // 旧会话记录的 value/revision 可能与新会话的实际配置不一致，残留会导致脏判断失真。
-  secretSavedStates.clear();
-  current = await fetchConfiguration();
-  bindAutosave();
-  render(current, themeController, backgroundController, userData);
-}
-
-function render(
-  snapshot: ConfigurationSnapshot,
-  themeController: ThemeController,
-  backgroundController: BackgroundController,
-  userData: UserDataController | null = null,
-): void {
-  current = snapshot;
-  renderSummary(snapshot);
-  renderThemeSelector(element("console-theme-selector"), themeController, backgroundController, userData);
-  renderPublicFields(snapshot);
-  renderSecretFields(snapshot);
-  bindTtsProviderState();
-  renderAgent(snapshot);
-  renderConfigurationNavigation();
-  bindRestart(snapshot);
-  bindValidation();
-}
-
-function renderSummary(snapshot: ConfigurationSnapshot): void {
-  const target = element("configuration-summary");
-  target.replaceChildren();
-  const invalid = snapshot.fields.filter((field) => !field.valid).length;
-  const pending = snapshot.fields.filter((field) => field.pendingRestart).length
-    + (snapshot.agent?.pendingRestart ? 1 : 0);
-  target.append(
-    badge(snapshot.fileExists ? "runtime.toml 已建立" : "runtime.toml 尚未建立", snapshot.fileExists ? "ok" : "warn"),
-    badge(invalid === 0 ? "本地预检通过" : "需要完成配置", invalid === 0 ? "ok" : "warn"),
-    badge(pending === 0 ? "无待重启变更" : `${pending} 项重启后生效`, pending === 0 ? "muted" : "warn"),
-  );
-}
-
-function renderPublicFields(snapshot: ConfigurationSnapshot): void {
-  const target = element("public-config-fields");
-  target.replaceChildren();
-  appendGroupedFields(
-    target,
-    snapshot.fields.filter((value) => value.sensitivity !== "secret"),
-    (field) => {
-    const row = document.createElement("div");
-    row.className = "config-row";
-    decorateTtsRow(row, field);
-    const label = document.createElement("label");
-    label.htmlFor = inputId(field.key);
-    label.textContent = configFieldLabel(field.key);
-    label.append(meta(field));
-    const input = fieldInput(field);
-    input.dataset.autosaveScope = "public";
-    row.append(label, input);
-    if (field.savedValue !== null && field.editable) {
-      const remove = button("恢复未保存值", "secondary");
-      remove.addEventListener("click", () => void removePublicField(field.key));
-      row.append(remove);
-    }
-      return row;
-    },
-    "runtime",
-  );
-  const save = element("save-public-config", HTMLButtonElement);
-  save.onclick = () => void savePublicFields();
-}
-
-function renderSecretFields(snapshot: ConfigurationSnapshot): void {
-  const target = element("secret-config-fields");
-  target.replaceChildren();
-  appendGroupedFields(
-    target,
-    snapshot.fields.filter((value) => value.sensitivity === "secret"),
-    (field) => {
-    const row = document.createElement("div");
-    row.className = "config-row secret-row";
-    decorateTtsRow(row, field);
-    const label = document.createElement("label");
-    label.htmlFor = inputId(field.key);
-    label.textContent = configFieldLabel(field.key);
-    label.append(meta(field));
-    const input = document.createElement("input");
-    input.id = inputId(field.key);
-    input.type = "password";
-    input.autocomplete = "new-password";
-    input.placeholder = field.configured ? "已配置；留空表示不修改" : "尚未配置";
-    input.disabled = !field.editable;
-    input.dataset.configKey = field.key;
-    input.dataset.autosaveScope = "secret";
-    const reveal = document.createElement("button");
-    reveal.type = "button";
-    reveal.className = "reveal-button";
-    reveal.textContent = "显示";
-    reveal.setAttribute("aria-pressed", "false");
-    reveal.setAttribute("aria-label", "显示或隐藏敏感值");
-    reveal.disabled = !field.editable;
-    reveal.addEventListener("click", () => togglePasswordReveal(reveal, input));
-    const wrap = document.createElement("div");
-    wrap.className = "password-field";
-    wrap.append(input, reveal);
-    const clearLabel = document.createElement("label");
-    clearLabel.className = "clear-secret";
-    const clear = document.createElement("input");
-    clear.type = "checkbox";
-    clear.dataset.clearKey = field.key;
-    clear.disabled = !field.editable || !field.configured;
-    clearLabel.append(clear, document.createTextNode(" 显式清除"));
-    row.append(label, wrap, clearLabel);
-      return row;
-    },
-    "secrets",
-  );
-  const save = element("save-secret-config", HTMLButtonElement);
-  save.onclick = () => void saveSecrets();
-}
-
-function renderAgent(snapshot: ConfigurationSnapshot): void {
-  const target = element("agent-config-fields");
-  target.replaceChildren();
-  const agent = snapshot.agent;
-  if (!agent || !agent.fileExists) {
-    target.textContent = "Agent 策略文件尚不可用；请检查默认 config/agent.toml 是否可写。";
-    element("save-agent-config", HTMLButtonElement).disabled = true;
-    return;
-  }
-  const documentValue = record(agent.savedValue);
-  const knowledge = record(documentValue.knowledge);
-  const embedding = record(knowledge.embedding);
-  const runningKnowledge = record(record(agent.runningValue).knowledge);
-  const runningEmbedding = record(runningKnowledge.embedding);
-  target.append(configurationGroup("memory-knowledge", "agent", fieldGroup("知识检索", [
-    selectField("知识检索模式", "agent-knowledge-mode", string(knowledge.mode) || "preflight", [
-      ["preflight", "preflight（高相关时条件注入）"],
-      ["tool", "tool（完全由 Agent 检索）"],
-      ["auto", "auto（紧急回退）"],
-    ], !agent.editable, "agent"),
-    checkboxField("本地语义召回", "agent-knowledge-embedding-enabled", embedding.enabled === true, !agent.editable, "agent"),
-    statusField(
-      `当前生效：${string(runningKnowledge.mode) || "preflight"} · 本地语义召回：${runningEmbedding.enabled === true ? "开启" : "关闭"}`,
-      `来源：${sourceLabel(agent.source)}${agent.pendingRestart ? " · 已保存变更等待重启" : ""}`,
-    ),
-    statusField(
-      "本地模型资源",
-      "首次开启会下载 BAAI/bge-small-zh-v1.5，并增加 CPU、内存占用；低配置服务器建议关闭。",
-    ),
-  ])));
-  const savedWebSearch = readAgentWebSearchConfig(documentValue);
-  const runningWebSearch = readAgentWebSearchConfig(agent.runningValue);
-  const tavilyKeyConfigured = snapshot.fields.some(
-    (field) => field.key === "tools.web_search.tavily.api_key" && field.configured,
-  );
-  const backendPendingRestart = savedWebSearch.backend !== runningWebSearch.backend;
-  const credentialStatus = statusField("Tavily 凭据", "");
-  credentialStatus.id = "agent-web-search-credential-status";
-  target.append(configurationGroup("online-tools", "agent", fieldGroup("联网搜索", [
-    statusField(
-      `当前生效后端：${webSearchBackendLabel(runningWebSearch.backend)}`,
-      `已保存后端：${webSearchBackendLabel(savedWebSearch.backend)} · ${backendPendingRestart ? "等待重启" : "当前已生效"}`,
-    ),
-    selectField("搜索后端", "agent-web-search-backend", savedWebSearch.backend, [
-      ["provider_native", "Provider 原生搜索"],
-      ["tavily", "Tavily"],
-      ["disabled", "关闭联网搜索"],
-    ], !agent.editable, "agent"),
-    numberField("Tavily 结果数", "agent-web-search-max-results", savedWebSearch.maxResults, 1, 10, !agent.editable, "agent"),
-    selectField("Tavily 搜索深度", "agent-web-search-depth", savedWebSearch.searchDepth, [
-      ["basic", "basic"],
-      ["advanced", "advanced"],
-    ], !agent.editable, "agent"),
-    selectField("Tavily 主题", "agent-web-search-topic", savedWebSearch.topic, [
-      ["general", "通用"],
-      ["news", "新闻"],
-      ["finance", "金融"],
-    ], !agent.editable, "agent"),
-    selectField("Tavily 时间范围", "agent-web-search-time-range", savedWebSearch.timeRange ?? "", [
-      ["", "不限"],
-      ["day", "最近一天"],
-      ["week", "最近一周"],
-      ["month", "最近一月"],
-      ["year", "最近一年"],
-    ], !agent.editable, "agent"),
-    numberField("连接超时（秒）", "agent-web-search-connect-timeout", savedWebSearch.connectTimeoutSeconds, 1, 3600, !agent.editable, "agent"),
-    numberField("首响应超时（秒）", "agent-web-search-first-response-timeout", savedWebSearch.firstResponseTimeoutSeconds, 1, 3600, !agent.editable, "agent"),
-    numberField("总超时（秒）", "agent-web-search-total-timeout", savedWebSearch.totalTimeoutSeconds, 1, 3600, !agent.editable, "agent"),
-    credentialStatus,
-  ])));
-  const backendSelect = element("agent-web-search-backend", HTMLSelectElement);
-  const refreshCredentialStatus = (): void => {
-    updateTavilyCredentialStatus(
-      isWebSearchBackend(backendSelect.value) ? backendSelect.value : "provider_native",
-      tavilyKeyConfigured,
-    );
-  };
-  backendSelect.addEventListener("change", refreshCredentialStatus);
-  refreshCredentialStatus();
-  target.append(configurationGroup("models-providers", "agent", renderOpenCodeProviders(
-    snapshot,
-    async (form) => {
-      let change: Record<string, unknown>;
-      try {
-        change = openCodeProviderChange(form);
-      } catch (cause) {
-        showResult(errorMessage(cause), true);
-        return;
-      }
-      await runSave(async () => updateAgentConfiguration(current!.agent!.revision, [change]));
-    },
-    async (id) => {
-      await runSave(async () => updateAgentConfiguration(current!.agent!.revision, [{
-        action: "remove_provider",
-        id,
-      }]));
-    },
-  )));
-  const openCodeKeyConfigured = snapshot.fields.some(
-    (field) => field.key === "provider.opencode.api_key" && field.configured,
-  );
-  const modelRoutes = record(documentValue.model_routes);
-  const routes = document.createElement("div");
-  routes.append(fieldGroup("模型路线", [
-    ...["private_main", "group_main", "aux"].map((routeName) => {
-      const route = record(modelRoutes[routeName]);
-      return textField(AGENT_ROUTE_LABELS[routeName] ?? routeName, `agent-route-${routeName}`, array(route.candidates).join(", "), !agent.editable, "agent");
-    }),
-    ...["private_search", "group_search"].map((routeName) => textField(
-      AGENT_ROUTE_LABELS[routeName] ?? routeName,
-      `agent-search-${routeName}`,
-      savedWebSearch.routes[routeName] ?? "",
-      !agent.editable,
-      "agent",
-    )),
-  ]));
-  routes.append(renderOpenCodeRouteHints(
-    !agent.editable,
-    readOpenCodeProviders(documentValue).filter((provider) => provider.enabled).map((provider) => provider.id),
-    openCodeKeyConfigured,
-  ));
-  target.append(configurationGroup("model-routing", "agent", routes));
-  const scenes = record(documentValue.scenes);
-  for (const sceneName of ["private", "group"]) {
-    const scene = record(scenes[sceneName]);
-    const row = document.createElement("div");
-    row.className = "config-row compact-row";
-    const label = document.createElement("label");
-    label.htmlFor = `agent-tool-${sceneName}`;
-    label.textContent = `${sceneName === "private" ? "私聊" : "群聊"} Tool Calling`;
-    const input = document.createElement("input");
-    input.id = `agent-tool-${sceneName}`;
-    input.type = "checkbox";
-    input.checked = scene.tool_calling_enabled === true;
-    input.disabled = !agent.editable;
-    input.dataset.autosaveScope = "agent-scene";
-    input.dataset.autosaveScene = sceneName;
-    row.append(label, input);
-    const sceneGroup = document.createElement("div");
-    sceneGroup.append(row);
-
-    const tools = document.createElement("fieldset");
-    tools.className = "tool-whitelist";
-    const legend = document.createElement("legend");
-    legend.textContent = `${sceneName === "private" ? "私聊" : "群聊"}工具白名单`;
-    tools.append(legend);
-    const savedNames = array(scene.enabled_tools).filter((value): value is string => typeof value === "string");
-    const visibleTools = agentToolOptions(snapshot.registeredTools, savedNames, agent.editable);
-    if (visibleTools.length === 0) {
-      const hint = document.createElement("p");
-      hint.className = "hint";
-      hint.textContent = "当前没有可用的已注册工具。";
-      tools.append(hint);
-    } else {
-      const grid = document.createElement("div");
-      grid.className = "tool-whitelist-grid";
-      for (const tool of visibleTools) {
-        grid.append(toolCheckbox(tool, sceneName));
-      }
-      tools.append(grid);
-    }
-    const saveScene = document.createElement("button");
-    saveScene.type = "button";
-    saveScene.className = "secondary tool-whitelist-save";
-    saveScene.textContent = `保存${sceneName === "private" ? "私聊" : "群聊"}配置`;
-    saveScene.disabled = !agent.editable;
-    saveScene.onclick = () => void saveAgentScene(sceneName);
-    tools.append(saveScene);
-    sceneGroup.append(tools);
-    target.append(configurationGroup("online-tools", "agent", sceneGroup));
-  }
-  const save = element("save-agent-config", HTMLButtonElement);
-  save.disabled = !agent.editable;
-  save.onclick = () => void saveAgent();
-}
-
-function appendGroupedFields(
-  target: HTMLElement,
-  fields: ConfigFieldSnapshot[],
-  row: (field: ConfigFieldSnapshot) => HTMLElement,
-  source: "runtime" | "secrets",
-): void {
-  const remaining = new Set(fields);
-  for (const business of BUSINESS_GROUPS) {
-    if (business.id === "advanced") continue;
-    const sections: HTMLElement[] = [];
-    for (const definition of FIELD_SECTIONS.filter((section) => section.business === business.id)) {
-      const grouped = fields.filter((field) => field.key.startsWith(definition.prefix));
-      if (grouped.length === 0) continue;
-      const label = source === "secrets" ? `${definition.label} · 敏感凭据` : definition.label;
-      sections.push(fieldGroup(label, grouped.map(row), source === "runtime" ? definition.description : undefined));
-      grouped.forEach((field) => remaining.delete(field));
-    }
-    if (sections.length > 0) target.append(configurationGroup(business.id, source, sections));
-  }
-  if (remaining.size > 0) {
-    const label = source === "secrets" ? "未归类敏感凭据" : "未归类受管字段";
-    target.append(configurationGroup("advanced", source, fieldGroup(label, [...remaining].map(row),
-      "当前前端尚未识别这些已登记字段；它们仍使用原配置键和原保存协议。")));
-  }
-}
-
-function configurationGroup(
-  group: ConfigurationBusinessGroup,
-  source: "runtime" | "secrets" | "agent" | "interface",
-  content: HTMLElement | HTMLElement[],
-): HTMLElement {
-  const wrapper = document.createElement("section");
-  wrapper.className = "configuration-content-group";
-  wrapper.dataset.configurationGroup = group;
-  wrapper.dataset.configurationSource = source;
-  wrapper.append(...(Array.isArray(content) ? content : [content]));
-  return wrapper;
-}
-
-function fieldGroup(label: string, rows: HTMLElement[], description?: string): HTMLElement {
-  const section = document.createElement("section");
-  section.className = "config-field-group";
-  const heading = document.createElement("h3");
-  heading.textContent = label;
-  const grid = document.createElement("div");
-  grid.className = "config-field-group-grid";
-  grid.append(...rows);
-  section.append(heading);
-  if (description) {
-    const hint = document.createElement("p");
-    hint.className = "config-field-group-hint";
-    hint.textContent = description;
-    section.append(hint);
-  }
-  section.append(grid);
-  return section;
-}
-
-function renderConfigurationNavigation(): void {
-  const tabs = element("configuration-business-tabs");
-  const content = element("configuration-business-content");
-  const description = element("configuration-business-description");
-  tabs.replaceChildren();
-  const groups = [...content.querySelectorAll<HTMLElement>("[data-configuration-group]")];
-  const availableIds = new Set(groups
-    .map((group) => group.dataset.configurationGroup)
-    .filter((group): group is ConfigurationBusinessGroup => isBusinessGroup(group)));
-  const available = BUSINESS_GROUPS.filter((item) => availableIds.has(item.id));
-  if (!availableIds.has(selectedBusinessGroup)) selectedBusinessGroup = available[0]?.id ?? "system-security";
-
-  for (const [index, item] of available.entries()) {
-    const tab = configurationTab(`configuration-business-${item.id}`, item.label, item.description);
-    tab.dataset.configurationGroup = item.id;
-    tab.setAttribute("aria-selected", String(item.id === selectedBusinessGroup));
-    tab.tabIndex = item.id === selectedBusinessGroup ? 0 : -1;
-    tab.addEventListener("click", () => {
-      selectedBusinessGroup = item.id;
-      renderConfigurationNavigation();
-    });
-    bindTabKeyboard(tab, tabs, index, available.length, () => {
-      selectedBusinessGroup = item.id;
-      renderConfigurationNavigation();
-    });
-    tabs.append(tab);
-  }
-  for (const group of groups) group.hidden = group.dataset.configurationGroup !== selectedBusinessGroup;
-  const selected = BUSINESS_GROUPS.find((item) => item.id === selectedBusinessGroup);
-  description.textContent = selected?.description ?? "";
-
-  const visibleSources = new Set(groups
-    .filter((group) => !group.hidden)
-    .map((group) => group.dataset.configurationSource));
-  element("save-public-config", HTMLButtonElement).hidden = !visibleSources.has("runtime");
-  element("save-secret-config", HTMLButtonElement).hidden = !visibleSources.has("secrets");
-  element("save-agent-config", HTMLButtonElement).hidden = !visibleSources.has("agent");
-  content.setAttribute("aria-labelledby", `configuration-business-${selectedBusinessGroup}`);
-}
-
-function isBusinessGroup(value: string | undefined): value is ConfigurationBusinessGroup {
-  return BUSINESS_GROUPS.some((group) => group.id === value);
-}
-
-/** 测试与页面共用同一份顺序，避免导航展示与字段归类出现两套事实来源。 */
-export function configurationBusinessGroups(keys: readonly string[]): ConfigurationBusinessGroup[] {
-  const available = new Set(keys.map(configurationBusinessGroup));
-  return BUSINESS_GROUPS.map((group) => group.id).filter((group) => available.has(group));
-}
-
-function configurationTab(id: string, label: string, description: string): HTMLButtonElement {
-  const tab = document.createElement("button");
-  tab.id = id;
-  tab.type = "button";
-  tab.className = "configuration-tab";
-  tab.setAttribute("role", "tab");
-  tab.setAttribute("aria-controls", "configuration-business-content");
-  tab.title = description;
-  tab.textContent = label;
-  return tab;
-}
-
-function bindTabKeyboard(
-  tab: HTMLButtonElement,
-  tablist: HTMLElement,
-  index: number,
-  count: number,
-  activate: () => void,
-): void {
-  tab.addEventListener("keydown", (event) => {
-    const tabs = [...tablist.querySelectorAll<HTMLButtonElement>("[role=tab]")];
-    let nextIndex: number | null = null;
-    if (event.key === "ArrowRight") nextIndex = (index + 1) % count;
-    if (event.key === "ArrowLeft") nextIndex = (index - 1 + count) % count;
-    if (event.key === "Home") nextIndex = 0;
-    if (event.key === "End") nextIndex = count - 1;
-    if (nextIndex !== null) {
-      event.preventDefault();
-      tabs[nextIndex]?.focus();
-      return;
-    }
-    if (event.key === "Enter" || event.key === " ") {
-      event.preventDefault();
-      activate();
-    }
-  });
-}
-
-export function publicConfigurationChanges(
-  fields: ConfigFieldSnapshot[],
-  values: ReadonlyMap<string, unknown>,
-): Array<Record<string, unknown>> {
-  const changes: Array<Record<string, unknown>> = [];
-  for (const field of fields.filter((value) => value.sensitivity === "public" && value.editable)) {
-    if (!values.has(field.key)) continue;
-    const value = values.get(field.key);
-    const baseline = field.savedValue ?? field.effectiveValue;
-    // 未配置的可选字段会显示为空输入框；用户未触碰时不能把空字符串误当成新配置提交。
-    if ((baseline === null || baseline === undefined) && isEmptyInputValue(value)) continue;
-    if (JSON.stringify(value) !== JSON.stringify(baseline)) {
-      changes.push({ action: "set", key: field.key, value });
-    }
-  }
-  return changes;
-}
-
-async function savePublicFields(): Promise<void> {
-  if (!current) return;
-  const values = new Map<string, unknown>();
-  for (const field of current.fields.filter((value) => value.sensitivity === "public" && value.editable)) {
-    const input = configInput(field.key);
-    if (!input.checkValidity()) {
-      input.reportValidity();
-      return showResult(`${configFieldLabel(field.key)}不符合页面输入范围，请修改后再保存。`, true);
-    }
-    try {
-      const value = ttsNumberRange(field.key)
-        ? parseTtsNumberValue(field.key, input.value)
-        : inputValue(input, field);
-      values.set(field.key, value);
-    } catch (cause) {
-      return showResult(errorMessage(cause), true);
-    }
-  }
-  const changes = publicConfigurationChanges(current.fields, values);
-  if (changes.length === 0) return showResult("没有需要保存的普通配置。", false);
-  await runSave(async () => updateRuntimeConfiguration(current!.revision, changes));
-}
-
-async function removePublicField(key: string): Promise<void> {
-  if (!current) return;
-  // 用户点击“恢复未保存值”即明确放弃该字段的页面输入：重建后该字段应显示服务端默认值，
-  // 因此把该输入排除在恢复集合之外。
-  await runSave(
-    async () => updateRuntimeConfiguration(current!.revision, [{ action: "remove", key }]),
-    new Set([`id:${inputId(key)}`]),
-  );
-}
-
-export function secretConfigurationChanges(
-  fields: ConfigFieldSnapshot[],
-  values: ReadonlyMap<string, string>,
-  clearKeys: ReadonlySet<string>,
-  dirtyKeys: ReadonlySet<string>,
-): Array<Record<string, unknown>> {
-  const changes: Array<Record<string, unknown>> = [];
-  for (const field of fields.filter((value) => value.sensitivity === "secret" && value.editable)) {
-    // 只提交真正发生变化的 Secret：已成功保存的 Secret 输入仍留在页面，但不能重复提交旧 revision。
-    if (!dirtyKeys.has(field.key)) continue;
-    if (clearKeys.has(field.key)) {
-      changes.push({ action: "clear", key: field.key, expected_revision: field.revision ?? "missing" });
-    } else {
-      const value = values.get(field.key) ?? "";
-      if (value.length > 0) {
-        changes.push({ action: "replace", key: field.key, value, expected_revision: field.revision ?? "missing" });
-      }
-    }
-  }
-  return changes;
-}
-
-async function saveSecrets(): Promise<void> {
-  if (!current) return;
-  // Secret 的脏状态、changes 与 expected_revision 必须在保存队列真正执行时重新计算：
-  // 前一个保存完成会更新 current snapshot 与 secretSavedStates，只有执行时才能拿到最新
-  // revision；若在入队前就计算，两个 Secret 连续失焦时第二个请求会重复携带第一个 Secret
-  // 的旧 revision，触发 config_conflict 并阻止第二个 Secret 保存。
-  let excluded: ReadonlySet<string> = EMPTY_EXCLUDED_KEYS;
-  await runSave(async () => {
-    const fields = current?.fields ?? [];
-    const values = new Map<string, string>();
-    const clearKeys = new Set<string>();
-    const dirtyKeys = new Set<string>();
-    for (const field of fields.filter((value) => value.sensitivity === "secret" && value.editable)) {
-      const value = element(inputId(field.key), HTMLInputElement).value;
-      const clear = document.querySelector<HTMLInputElement>(`input[data-clear-key="${field.key}"]`);
-      const clearChecked = clear?.checked === true;
-      values.set(field.key, value);
-      if (secretIsDirty(field, value, clearChecked)) {
-        dirtyKeys.add(field.key);
-        if (clearChecked) clearKeys.add(field.key);
-      }
-    }
-    const changes = secretConfigurationChanges(fields, values, clearKeys, dirtyKeys);
-    if (changes.length === 0) {
-      showResult("留空不会清除 secret；当前没有显式变更。", false);
-      return null;
-    }
-    // 显式清除成功后，输入框和“显式清除”勾选都应复位到服务端状态，不能通过重建后的恢复保留旧值。
-    const nextExcluded = new Set<string>();
-    for (const key of clearKeys) {
-      nextExcluded.add(`clear:${key}`);
-      nextExcluded.add(`id:${inputId(key)}`);
-    }
-    excluded = nextExcluded;
-    const snapshot = await updateSecretConfiguration(changes);
-    rememberSecretSavedStates(snapshot, dirtyKeys, values, clearKeys);
-    return snapshot;
-  }, () => excluded);
-}
-
-function secretIsDirty(field: ConfigFieldSnapshot, value: string, clearChecked: boolean): boolean {
-  const saved = secretSavedStates.get(field.key);
-  if (saved) {
-    if (clearChecked !== saved.clear) return true;
-    return !clearChecked && value.length > 0 && value !== saved.value;
-  }
-  // 首次交互沿用既有语义：显式清除只在已配置时生效；非空输入视为新值；留空表示不修改。
-  if (clearChecked) return field.configured;
-  return value.length > 0;
-}
-
-/** 保存成功后按提交结果记录每个 Secret 的最新已保存状态，供后续脏判断使用。 */
-function rememberSecretSavedStates(
-  snapshot: ConfigurationSnapshot,
-  dirtyKeys: ReadonlySet<string>,
-  values: ReadonlyMap<string, string>,
-  clearKeys: ReadonlySet<string>,
-): void {
-  for (const key of dirtyKeys) {
-    const field = snapshot.fields.find((candidate) => candidate.key === key);
-    if (!field) continue;
-    secretSavedStates.set(key, {
-      value: clearKeys.has(key) ? "" : values.get(key) ?? "",
-      clear: clearKeys.has(key),
-      revision: field.revision ?? null,
-    });
-  }
-}
-
-async function saveAgent(): Promise<void> {
-  if (!current?.agent) return;
-  let webSearchChange: Record<string, unknown>;
-  try {
-    webSearchChange = webSearchConfigChange(webSearchFormConfig());
-  } catch (cause) {
-    showResult(errorMessage(cause), true);
-    return;
-  }
-  const documentValue = record(current.agent.savedValue);
-  const scenes = record(documentValue.scenes);
-  const embedding = record(record(documentValue.knowledge).embedding);
-  const changes: unknown[] = [{
-    action: "set_knowledge",
-    mode: element("agent-knowledge-mode", HTMLSelectElement).value,
-    embedding: {
-      enabled: element("agent-knowledge-embedding-enabled", HTMLInputElement).checked,
-      cache_dir: string(embedding.cache_dir) || "cache/knowledge-embedding",
-    },
-  }, webSearchChange];
-  for (const routeName of ["private_main", "group_main", "aux"]) {
-    const candidates = element(`agent-route-${routeName}`, HTMLInputElement).value
-      .split(",").map((value) => value.trim()).filter(Boolean);
-    changes.push({ action: "set_model_route", name: routeName, candidates });
-  }
-  const savedRoutes = readAgentWebSearchConfig(documentValue).routes;
-  changes.push(...webSearchRouteChanges(savedRoutes, {
-    private_search: element("agent-search-private_search", HTMLInputElement).value,
-    group_search: element("agent-search-group_search", HTMLInputElement).value,
-  }));
-  for (const sceneName of ["private", "group"]) {
-    changes.push({ action: "set_scene", scene: sceneName, config: agentSceneConfig(sceneName, scenes) });
-  }
-  await runSave(async () => updateAgentConfiguration(current!.agent!.revision, changes));
-}
-
-async function saveAgentScene(sceneName: string): Promise<void> {
-  if (!current?.agent) return;
-  const scenes = record(record(current.agent.savedValue).scenes);
-  await runSave(async () => updateAgentConfiguration(current!.agent!.revision, [{
-    action: "set_scene",
-    scene: sceneName,
-    config: agentSceneConfig(sceneName, scenes),
-  }]));
-}
-
-async function saveOpenCodeProvider(id: string): Promise<void> {
-  if (!current?.agent) return;
-  const baseUrl = element(`${id}-base-url`, HTMLInputElement);
-  const timeout = element(`${id}-timeout`, HTMLInputElement);
-  const saved = readOpenCodeProviders(current.agent.savedValue).find((provider) => provider.id === id);
-  // 未添加的预设只能通过“添加 Provider”显式启用，浏览默认字段不能改变配置。
-  if (!saved?.enabled) return;
-  const form = {
-    ...saved,
-    baseUrl: baseUrl.value,
-    requestTimeoutSeconds: timeout.value.trim() ? Number(timeout.value) : null,
-    enabled: true,
-  };
-  if (!shouldAutosaveOnBlur({ scope: "agent", value: form, baseline: saved })) return;
-  let change: Record<string, unknown>;
-  try {
-    change = openCodeProviderChange(form);
-  } catch (cause) {
-    showResult(errorMessage(cause), true);
-    return;
-  }
-  await runSave(async () => updateAgentConfiguration(current?.agent?.revision ?? "missing", [change]));
-}
-
-function agentSceneConfig(sceneName: string, scenes: Record<string, unknown>): Record<string, unknown> {
-  const toolInputs = document.querySelectorAll<HTMLInputElement>(`input[data-agent-tool="${sceneName}"]`);
-  return {
-    ...record(scenes[sceneName]),
-    tool_calling_enabled: element(`agent-tool-${sceneName}`, HTMLInputElement).checked,
-    enabled_tools: selectedAgentToolNames(toolInputs),
-  };
-}
-
-function webSearchFormConfig(): AgentWebSearchConfig {
-  const backend = element("agent-web-search-backend", HTMLSelectElement).value;
-  const searchDepth = element("agent-web-search-depth", HTMLSelectElement).value;
-  const topic = element("agent-web-search-topic", HTMLSelectElement).value;
-  const timeRange = element("agent-web-search-time-range", HTMLSelectElement).value;
-  if (!isWebSearchBackend(backend)) throw new Error("联网搜索后端无效");
-  return {
-    backend,
-    maxResults: integerInput("agent-web-search-max-results"),
-    searchDepth: searchDepth === "advanced" ? "advanced" : "basic",
-    topic: topic === "news" || topic === "finance" ? topic : "general",
-    timeRange: isWebSearchTimeRange(timeRange) ? timeRange : null,
-    connectTimeoutSeconds: integerInput("agent-web-search-connect-timeout"),
-    firstResponseTimeoutSeconds: integerInput("agent-web-search-first-response-timeout"),
-    totalTimeoutSeconds: integerInput("agent-web-search-total-timeout"),
-    routes: {},
-  };
-}
-
-function bindAutosave(): void {
-  if (autosaveBound) return;
-  autosaveBound = true;
-  document.addEventListener("focusout", (event) => {
-    const target = event.target;
-    if (!(target instanceof HTMLInputElement) && !(target instanceof HTMLSelectElement)) return;
-    const related = event.relatedTarget;
-    queuedFocusRestoreId = related instanceof HTMLElement && related.id ? related.id : null;
-    // 点击显式保存按钮会先触发输入框 blur；此时由按钮提交，避免同一 revision 入队两次。
-    if (related instanceof HTMLElement && shouldDeferAutosaveToButton(target, related)) return;
-    void autosaveBlur(target);
-  });
-}
-
-function shouldDeferAutosaveToButton(
-  target: HTMLInputElement | HTMLSelectElement,
-  related: HTMLElement,
-): boolean {
-  if (related.id === "save-secret-config") {
-    return target.dataset.autosaveScope === "secret" || target.dataset.clearKey !== undefined;
-  }
-  if (related.id === "save-public-config") return target.dataset.autosaveScope === "public";
-  if (related.id === "save-agent-config") return target.dataset.autosaveScope === "agent";
-  return target.dataset.autosaveProvider !== undefined && related.matches(".provider-action");
-}
-
-async function autosaveBlur(target: HTMLInputElement | HTMLSelectElement): Promise<void> {
-  if (target.disabled || !current) return;
-  const scene = target.dataset.autosaveScene;
-  if (target.dataset.configKey) {
-    const field = current.fields.find((value) => value.key === target.dataset.configKey);
-    if (!field || !field.editable) return;
-    if (field.sensitivity === "secret") {
-      const clear = document.querySelector<HTMLInputElement>(`input[data-clear-key="${field.key}"]`);
-      if (!secretIsDirty(field, target.value, clear?.checked === true)) return;
-      await saveSecrets();
-      return;
-    }
-    const value = inputValue(target, field);
-    if (!shouldAutosaveOnBlur({ scope: "public", value, baseline: field.savedValue ?? field.effectiveValue })) return;
-    await savePublicFields();
-    return;
-  }
-  if (target.dataset.clearKey) {
-    const field = current.fields.find((value) => value.key === target.dataset.clearKey);
-    if (!(target instanceof HTMLInputElement) || !field || !field.editable || !target.checked) return;
-    if (!secretIsDirty(field, element(inputId(field.key), HTMLInputElement).value, true)) return;
-    await saveSecrets();
-    return;
-  }
-  const providerId = target.dataset.autosaveProvider;
-  if (providerId) {
-    await saveOpenCodeProvider(providerId);
-    return;
-  }
-  if (scene) {
-    if (agentSceneChanged(scene)) await saveAgentScene(scene);
-    return;
-  }
-  if (target.dataset.autosaveScope === "agent" && agentFieldChanged(target.id)) await saveAgent();
-}
-
-function agentSceneChanged(sceneName: string): boolean {
-  if (!current?.agent) return false;
-  const savedScenes = record(record(current.agent.savedValue).scenes);
-  return shouldAutosaveOnBlur({
-    scope: "agent",
-    value: agentSceneConfig(sceneName, savedScenes),
-    baseline: record(savedScenes[sceneName]),
-  });
-}
-
-function agentFieldChanged(id: string): boolean {
-  if (!current?.agent) return false;
-  const saved = record(current.agent.savedValue);
-  const webSearch = readAgentWebSearchConfig(current.agent.savedValue);
-  const currentValue = id === "agent-knowledge-mode" ? element(id, HTMLSelectElement).value
-    : id === "agent-knowledge-embedding-enabled" ? element(id, HTMLInputElement).checked
-    : id.startsWith("agent-web-search-") ? agentWebSearchInputValue(id)
-    : id.startsWith("agent-route-") ? element(id, HTMLInputElement).value.split(",").map((value) => value.trim()).filter(Boolean)
-    : id.startsWith("agent-search-") ? element(id, HTMLInputElement).value.trim()
-    : null;
-  const baseline = id === "agent-knowledge-mode" ? string(record(saved.knowledge).mode) || "preflight"
-    : id === "agent-knowledge-embedding-enabled" ? record(record(saved.knowledge).embedding).enabled === true
-    : id.startsWith("agent-web-search-") ? webSearch[agentWebSearchKey(id)]
-    : id.startsWith("agent-route-") ? array(record(record(saved.model_routes)[id.replace("agent-route-", "")]).candidates).map(string)
-    : id.startsWith("agent-search-") ? webSearch.routes[id.replace("agent-search-", "")]
-    : null;
-  return currentValue !== null && shouldAutosaveOnBlur({ scope: "agent", value: currentValue, baseline });
-}
-
-function agentWebSearchKey(id: string): keyof AgentWebSearchConfig {
-  return ({
-    "agent-web-search-backend": "backend",
-    "agent-web-search-max-results": "maxResults",
-    "agent-web-search-depth": "searchDepth",
-    "agent-web-search-topic": "topic",
-    "agent-web-search-time-range": "timeRange",
-    "agent-web-search-connect-timeout": "connectTimeoutSeconds",
-    "agent-web-search-first-response-timeout": "firstResponseTimeoutSeconds",
-    "agent-web-search-total-timeout": "totalTimeoutSeconds",
-  } as const)[id] ?? "backend";
-}
-
-function agentWebSearchInputValue(id: string): unknown {
-  const key = agentWebSearchKey(id);
-  if (key === "backend" || key === "searchDepth" || key === "topic" || key === "timeRange") {
-    return element(id, HTMLSelectElement).value;
-  }
-  return Number(element(id, HTMLInputElement).value);
-}
-
-function toolCheckbox(tool: AgentToolOption, sceneName: string): HTMLElement {
-  const label = document.createElement("label");
-  label.className = "tool-checkbox";
-  label.title = tool.description;
-  const input = document.createElement("input");
-  input.type = "checkbox";
-  input.value = tool.name;
-  input.checked = tool.checked;
-  input.disabled = tool.disabled;
-  input.dataset.agentTool = sceneName;
-  input.dataset.autosaveScene = sceneName;
-  const name = document.createElement("span");
-  name.textContent = tool.name === "image_generation" ? "图片生成" : tool.name;
-  label.append(input, name);
-  if (!tool.registered) {
-    const state = document.createElement("span");
-    state.className = "tool-registration-state";
-    state.textContent = "当前进程未注册";
-    label.append(state);
-  }
-  return label;
-}
-
-function bindRestart(snapshot: ConfigurationSnapshot): void {
-  const restart = element("restart-service", HTMLButtonElement);
-  restart.disabled = !snapshot.restartAvailable;
-  restart.title = snapshot.restartAvailable ? "通过当前运行目录的 botctl 重启" : "当前运行目录没有可用的 botctl 重启脚本";
-  restart.onclick = async () => {
-    if (!window.confirm("确定要重启服务吗？控制台会短暂离线。")) return;
-    restart.disabled = true;
-    try {
-      showResult(await requestRestart(), false);
-    } catch (cause) {
-      showResult(errorMessage(cause), true);
-      restart.disabled = !snapshot.restartAvailable;
-    }
-  };
-}
-
-function bindValidation(): void {
-  element("validate-config", HTMLButtonElement).onclick = async () => {
-    try {
-      const result = await validateConfiguration();
-      showResult(result.message, !result.valid);
-    } catch (cause) {
-      showResult(errorMessage(cause), true);
-    }
-  };
-}
-
-async function runSave(
-  action: () => Promise<ConfigurationSnapshot | null>,
-  excludeKeys: ReadonlySet<string> | (() => ReadonlySet<string>) = EMPTY_EXCLUDED_KEYS,
-): Promise<ConfigurationSnapshot | null> {
-  const save = async (): Promise<ConfigurationSnapshot | null> => {
-    setButtonsDisabled(true);
-    try {
-      const snapshot = await action();
-      if (!snapshot) return null;
-      if (!currentThemeController || !currentBackgroundController) throw new Error("界面控制器尚未初始化");
-      const restoreId = queuedFocusRestoreId;
-      queuedFocusRestoreId = null;
-      // 保存完成后会重建配置 DOM；重建前先完整收集所有 input/select 的当前值
-      // （含 checkbox、select 与 Secret 输入），重建后恢复，避免覆盖其他字段未保存的输入。
-      const captured = captureConfigurationInputState();
-      render(snapshot, currentThemeController, currentBackgroundController, currentUserDataController);
-      const restoredExcluded = typeof excludeKeys === "function" ? excludeKeys() : excludeKeys;
-      restoreConfigurationInputState(captured, restoredExcluded);
-      if (restoreId) document.getElementById(restoreId)?.focus();
-      showResult("配置已真实持久化；标记为“重启后生效”的项需按部署方式重启服务。", false);
-      return snapshot;
-    } catch (cause) {
-      if (cause instanceof ConsoleApiError && cause.code === "config_conflict") {
-        showResult("配置文件已被其他操作修改。请刷新后重新合并，旧 revision 未覆盖新文件。", true);
-      } else {
-        showResult(errorMessage(cause), true);
-      }
-      return null;
-    } finally {
-      setButtonsDisabled(false);
-    }
-  };
-  saveQueue = saveQueue.then(save, save);
-  return saveQueue;
-}
-
-const CONFIGURATION_ROOT_ID = "configuration";
-
-/** 在重建配置 DOM 前收集所有用户可编辑输入的当前状态，键由 `inputCaptureKey` 稳定生成。 */
-function captureConfigurationInputState(): Map<string, string | boolean> {
-  const captured = new Map<string, string | boolean>();
-  const root = document.getElementById(CONFIGURATION_ROOT_ID);
-  if (!root) return captured;
-  for (const input of root.querySelectorAll<HTMLInputElement | HTMLSelectElement>("input, select")) {
-    if (input instanceof HTMLInputElement && input.type === "file") continue;
-    const key = inputCaptureKey(input);
-    if (!key) continue;
-    captured.set(key, input instanceof HTMLSelectElement || input.type !== "checkbox" ? input.value : input.checked);
-  }
-  return captured;
-}
-
-function inputCaptureKey(input: HTMLInputElement | HTMLSelectElement): string | null {
-  if (input.id) return `id:${input.id}`;
-  if (input.dataset.clearKey) return `clear:${input.dataset.clearKey}`;
-  if (input instanceof HTMLInputElement && input.type === "checkbox") {
-    if (input.dataset.agentTool) return `tool:${input.dataset.agentTool}:${input.value}`;
-    const name = input.getAttribute("name");
-    if (name === "console-theme" || name === "console-background") return `${name}:${input.value}`;
-  }
-  return null;
-}
-
-/**
- * 重建后恢复输入状态。`excludeKeys` 中的键保持重建后的服务端快照值（例如显式清除的
- * Secret、用户主动“恢复未保存值”的字段），避免把用户已决定丢弃的旧值再写回页面。
- */
-function restoreConfigurationInputState(
-  captured: ReadonlyMap<string, string | boolean>,
-  excludeKeys: ReadonlySet<string>,
-): void {
-  const root = document.getElementById(CONFIGURATION_ROOT_ID);
-  if (!root) return;
-  for (const input of root.querySelectorAll<HTMLInputElement | HTMLSelectElement>("input, select")) {
-    if (input instanceof HTMLInputElement && input.type === "file") continue;
-    const key = inputCaptureKey(input);
-    if (!key || excludeKeys.has(key) || !captured.has(key)) continue;
-    const value = captured.get(key);
-    if (value === undefined) continue;
-    if (input instanceof HTMLSelectElement || input.type !== "checkbox") {
-      // select 只恢复仍然存在的选项；选项列表随数据变化时以重建后的快照为准。
-      if (input instanceof HTMLSelectElement && ![...input.options].some((option) => option.value === value)) continue;
-      input.value = String(value);
-    } else {
-      input.checked = value === true;
-    }
-  }
-}
-
-function fieldInput(field: ConfigFieldSnapshot): HTMLInputElement | HTMLSelectElement {
-  const value = field.savedValue ?? field.effectiveValue;
-  if (field.key === "delivery.tts.provider") {
-    const select = document.createElement("select");
-    select.id = inputId(field.key);
-    select.dataset.configKey = field.key;
-    select.disabled = !field.editable;
-    const currentValue = value === null || value === undefined ? "disabled" : String(value);
-    for (const [optionValue, label] of ttsProviderOptions(currentValue)) {
-      const option = document.createElement("option");
-      option.value = optionValue;
-      option.textContent = label;
-      select.append(option);
-    }
-    select.value = currentValue;
-    return select;
-  }
-  if (field.key === "command.prefix") {
-    const select = document.createElement("select");
-    select.id = inputId(field.key);
-    select.dataset.configKey = field.key;
-    select.disabled = !field.editable;
-    const currentValue = value === null || value === undefined ? "/" : String(value);
-    const options: Array<[string, string]> = [
-      ["/", "/（默认）"],
-      ["#", "#"],
-      ["*", "*"],
-    ];
-    if (!options.some(([option]) => option === currentValue)) {
-      options.push([currentValue, `${currentValue}（当前自定义值）`]);
-    }
-    for (const [optionValue, label] of options) {
-      const option = document.createElement("option");
-      option.value = optionValue;
-      option.textContent = label;
-      select.append(option);
-    }
-    select.value = currentValue;
-    return select;
-  }
-  const input = document.createElement("input");
-  input.id = inputId(field.key);
-  input.dataset.configKey = field.key;
-  input.disabled = !field.editable;
-  if (field.valueType === "boolean") {
-    input.type = "checkbox";
-    input.checked = value === true;
-  } else {
-    input.type = field.valueType === "integer" ? "number" : "text";
-    input.value = Array.isArray(value) ? value.join(", ") : value === null || value === undefined ? "" : String(value);
-    const range = ttsNumberRange(field.key);
-    if (range) {
-      input.min = String(range[0]);
-      input.max = String(range[1]);
-      input.step = "1";
-      input.required = true;
-    }
-  }
-  return input;
-}
-
-function decorateTtsRow(row: HTMLElement, field: ConfigFieldSnapshot): void {
-  row.dataset.configFieldKey = field.key;
-  if (field.key.startsWith("delivery.tts.qwen_")) row.dataset.ttsQwenField = "true";
-}
-
-/** 关闭 Provider 只做视觉提示，字段始终保持可编辑且不会生成清除操作。 */
-function bindTtsProviderState(): void {
-  const provider = document.getElementById(inputId("delivery.tts.provider"));
-  if (!(provider instanceof HTMLSelectElement)) return;
-  const refresh = (): void => {
-    const disabled = provider.value === "disabled";
-    for (const row of document.querySelectorAll<HTMLElement>("[data-tts-qwen-field='true']")) {
-      row.classList.toggle("config-row-muted", disabled);
-    }
-  };
-  provider.addEventListener("change", refresh);
-  refresh();
-}
-
-function inputValue(input: HTMLInputElement | HTMLSelectElement, field: ConfigFieldSnapshot): unknown {
-  if (field.valueType === "boolean") return input instanceof HTMLInputElement && input.checked;
-  if (field.valueType === "integer") return Number.parseInt(input.value, 10);
-  if (field.valueType === "string_list") return input.value.split(",").map((value) => value.trim()).filter(Boolean);
-  return input.value.trim();
-}
-
-function configInput(key: string): HTMLInputElement | HTMLSelectElement {
-  const value = document.getElementById(inputId(key));
-  if (!(value instanceof HTMLInputElement) && !(value instanceof HTMLSelectElement)) {
-    throw new Error(`缺少配置输入 #${inputId(key)}`);
-  }
-  return value;
-}
-
-function isEmptyInputValue(value: unknown): boolean {
-  return value === "" || (Array.isArray(value) && value.length === 0);
-}
-
-function meta(field: ConfigFieldSnapshot): HTMLElement {
-  const value = document.createElement("span");
-  value.className = "field-meta";
-  const flags = [sourceLabel(field.source), field.applyMode === "restart" ? "重启后生效" : "立即生效"];
-  if (field.overridden) flags.push("已覆盖 .env");
-  if (field.pendingRestart) flags.push("等待重启");
-  if (!field.editable) flags.push("只读");
-  value.textContent = flags.join(" · ");
-  return value;
-}
-
-function sourceLabel(source: string): string {
-  return ({ environment: "环境变量", managed_toml: "runtime.toml", agent_toml: "agent.toml", encrypted_secret: "加密存储", default: "默认值", not_configured: "未配置" } as Record<string, string>)[source] ?? source;
-}
-
-function textField(labelText: string, id: string, value: string, disabled: boolean, autosaveScope?: AutosaveScope): HTMLElement {
-  const row = document.createElement("div");
-  row.className = "config-row";
-  const label = document.createElement("label");
-  label.htmlFor = id;
-  label.textContent = labelText;
-  const input = document.createElement("input");
-  input.id = id;
-  input.type = "text";
-  input.value = value;
-  input.disabled = disabled;
-  if (autosaveScope) input.dataset.autosaveScope = autosaveScope;
-  row.append(label, input);
-  return row;
-}
-
-function numberField(
-  labelText: string,
-  id: string,
-  value: number,
-  min: number,
-  max: number,
-  disabled: boolean,
-  autosaveScope?: AutosaveScope,
-): HTMLElement {
-  const row = document.createElement("div");
-  row.className = "config-row";
-  const label = document.createElement("label");
-  label.htmlFor = id;
-  label.textContent = labelText;
-  const input = document.createElement("input");
-  input.id = id;
-  input.type = "number";
-  input.min = String(min);
-  input.max = String(max);
-  input.step = "1";
-  input.value = String(value);
-  input.disabled = disabled;
-  if (autosaveScope) input.dataset.autosaveScope = autosaveScope;
-  row.append(label, input);
-  return row;
-}
-
-function selectField(
-  labelText: string,
-  id: string,
-  value: string,
-  options: Array<[string, string]>,
-  disabled: boolean,
-  autosaveScope?: AutosaveScope,
-): HTMLElement {
-  const row = document.createElement("div");
-  row.className = "config-row";
-  const label = document.createElement("label");
-  label.htmlFor = id;
-  label.textContent = labelText;
-  const select = document.createElement("select");
-  select.id = id;
-  select.disabled = disabled;
-  if (autosaveScope) select.dataset.autosaveScope = autosaveScope;
-  for (const [optionValue, optionLabel] of options) {
-    const option = document.createElement("option");
-    option.value = optionValue;
-    option.textContent = optionLabel;
-    select.append(option);
-  }
-  select.value = value;
-  row.append(label, select);
-  return row;
-}
-
-function checkboxField(labelText: string, id: string, checked: boolean, disabled: boolean, autosaveScope?: AutosaveScope): HTMLElement {
-  const row = document.createElement("div");
-  row.className = "config-row compact-row";
-  const label = document.createElement("label");
-  label.htmlFor = id;
-  label.textContent = labelText;
-  const input = document.createElement("input");
-  input.id = id;
-  input.type = "checkbox";
-  input.checked = checked;
-  input.disabled = disabled;
-  if (autosaveScope) input.dataset.autosaveScope = autosaveScope;
-  row.append(label, input);
-  return row;
-}
-
-function statusField(summary: string, detail: string): HTMLElement {
-  const row = document.createElement("div");
-  row.className = "config-row";
-  const label = document.createElement("strong");
-  label.textContent = summary;
-  const meta = document.createElement("span");
-  meta.className = "field-meta";
-  meta.textContent = detail;
-  row.append(label, meta);
-  return row;
-}
-
-function updateTavilyCredentialStatus(backend: AgentWebSearchBackend, configured: boolean): void {
-  const row = element("agent-web-search-credential-status");
-  const summary = row.querySelector("strong");
-  const detail = row.querySelector(".field-meta");
-  if (!summary || !detail) throw new Error("缺少 Tavily 凭据状态元素");
-  const notice = tavilyCredentialNotice(backend, configured);
-  summary.textContent = configured ? "Tavily API Key：已配置" : "Tavily API Key：未配置";
-  detail.textContent = notice || (configured
-    ? "密钥保存在安全配置中心，不会写入 agent.toml 或回传浏览器。"
-    : "可在“联网与工具”中配置；未选择 Tavily 时不影响其他搜索后端。");
-  row.classList.toggle("config-row-warning", notice.length > 0);
-}
-
-function badge(text: string, kind: string): HTMLElement {
-  const value = document.createElement("span");
-  value.className = `config-badge config-badge-${kind}`;
-  value.textContent = text;
-  return value;
-}
-
-function button(text: string, kind: string): HTMLButtonElement {
-  const value = document.createElement("button");
-  value.type = "button";
-  value.className = kind;
-  value.textContent = text;
-  return value;
-}
-
-function inputId(key: string): string { return `config-${key.replaceAll(".", "-")}`; }
-function record(value: unknown): Record<string, unknown> { return typeof value === "object" && value !== null && !Array.isArray(value) ? value as Record<string, unknown> : {}; }
-function array(value: unknown): unknown[] { return Array.isArray(value) ? value : []; }
-function string(value: unknown): string { return typeof value === "string" ? value : ""; }
-function positiveNumber(value: unknown, fallback: number): number { return typeof value === "number" && Number.isFinite(value) && value > 0 ? value : fallback; }
-function integerInput(id: string): number { return Number(element(id, HTMLInputElement).value); }
-function isWebSearchBackend(value: string): value is AgentWebSearchBackend { return value === "provider_native" || value === "tavily" || value === "disabled"; }
-function isWebSearchTimeRange(value: string): value is NonNullable<AgentWebSearchConfig["timeRange"]> { return value === "day" || value === "week" || value === "month" || value === "year"; }
-function webSearchBackendLabel(value: AgentWebSearchBackend): string {
-  return ({ provider_native: "Provider 原生搜索", tavily: "Tavily", disabled: "已关闭" } as const)[value];
-}
-
-function showResult(message: string, error: boolean): void {
-  const target = element("configuration-result");
-  target.textContent = message;
-  target.className = error ? "error" : "success";
-  showToast(message, error);
-}
-
-/** 右上角浮层提醒；进行中的消息不设置自动隐藏，避免转圈提示被提前关掉。 */
-function showToast(message: string, error: boolean): void {
-  const toast = element("console-toast");
-  toast.textContent = message;
-  toast.className = `console-toast ${error ? "console-toast-error" : "console-toast-success"}`;
-  toast.hidden = false;
-  if (toastTimer !== undefined) window.clearTimeout(toastTimer);
-  if (!message.startsWith("正在")) {
-    toastTimer = window.setTimeout(() => {
-      toast.hidden = true;
-      toastTimer = undefined;
-    }, 8_000);
-  }
-}
-
-function errorMessage(cause: unknown): string { return cause instanceof Error ? cause.message : "配置操作失败"; }
-
-function setButtonsDisabled(disabled: boolean): void {
-  for (const id of ["save-public-config", "save-secret-config", "save-agent-config", "validate-config"]) {
-    element(id, HTMLButtonElement).disabled = disabled;
-  }
-  for (const button of document.querySelectorAll<HTMLButtonElement>(".tool-whitelist-save")) {
-    button.disabled = disabled || current?.agent?.editable !== true;
-  }
-  for (const button of document.querySelectorAll<HTMLButtonElement>(".provider-action")) {
-    button.disabled = disabled || current?.agent?.editable !== true;
-  }
-}
-
-function element<T extends HTMLElement>(id: string, constructor?: { new(): T }): T {
-  const value = document.getElementById(id);
-  if (!value || (constructor && !(value instanceof constructor))) throw new Error(`缺少页面元素 #${id}`);
-  return value as T;
-}
